Accept fractional hours in every stage, not just the collector
x_collector's --hours is a float and the README documents `--hours 6.3`
for targeting the 6.25h NSE cash session, but the two stages downstream of
it still took an int. run_pipeline.sh forwards the same $HOURS to
whichever collector COLLECTOR selects and then to the processing stage, so
a fractional window died on "invalid int value: '6.3'" -- at stage 1 for
COLLECTOR=nitter, and at stage 2 for the default x.com path, which is the
worse of the two because collection has already run for 30-45 minutes by
then. The raw JSONL survives, so it is recoverable by re-running stage 2
by hand, but nothing says so at the point it fails.

Make the whole chain float: the config field both collectors take their
--hours default from, the Nitter collector's --hours, and the processing
stage's --lookback-hours. The arithmetic already worked -- timedelta takes
floats -- so this is only about what argparse and pydantic will accept.

Split build_parser() out of main() in both stages so the accepted argument
types are testable without launching a browser or running a full
processing pass. The defect was entirely in the type declarations, which
no test could reach while the parsers were buried in main().

+++ src/utils/config_loader.py
@@ -42,7 +42,11 @@ class ScraperConfig(BaseModel):
     related_hashtags: list[str] = []
     search_path_template: str
     nitter_hosts: list[str]
-    hours_lookback: int
+    # Fractional, because a collection window is not always a whole number of hours: the
+    # NSE cash session is 6.25h, and --offset-hours targets it directly. Every stage that
+    # consumes this has to agree, or a run that collects 6.3h of tweets dies in the next
+    # stage on "invalid int value: '6.3'".
+    hours_lookback: float
     min_tweets_target: int
     headless: bool
     worker_pool_size: int
